+++ src/test/java/nextstep/subway/acceptance/LineApiRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line;
+package nextstep.subway.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
