+++ src/test/java/nextstep/subway/acceptance/SectionApiRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section;
+package nextstep.subway.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
